refactor : move matching service package QUZ-83
