+++ matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/domain/game/GameVO.kt
@@ -1,0 +1,3 @@
+package com.grepp.quizy.matching.domain.game
+
+@JvmInline value class GameRoomId(val value: Long)
