+++ matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/domain/match/MatchingConstants.kt
@@ -1,0 +1,3 @@
+package com.grepp.quizy.matching.domain.match
+
+const val MATCHING_K = 5
