+++ matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/domain/match/MatchingEvents.kt
@@ -1,0 +1,7 @@
+package com.grepp.quizy.matching.domain.match
+
+import com.grepp.quizy.matching.domain.user.UserId
+
+data class UserWaitingRegisteredEvent(val userId: UserId)
+
+data class PersonalMatchingSucceedEvent(val userId: Long, val gameRoomId: Long, val totalIds: List<Long>)
