+++ matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/domain/match/MatchingPoolRepository.kt
@@ -1,0 +1,12 @@
+package com.grepp.quizy.matching.domain.match
+
+import com.grepp.quizy.matching.domain.user.UserId
+import com.grepp.quizy.matching.domain.user.UserVector
+
+interface MatchingPoolRepository {
+    fun saveVector(userId: UserId, userVector: UserVector)
+
+    fun findNearestUser(userStatus: UserStatus): List<UserStatus>
+
+    fun remove(userId: UserId)
+}
