+++ matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/domain/match/MatchingQueueRepository.kt
@@ -1,0 +1,15 @@
+package com.grepp.quizy.matching.domain.match
+
+import com.grepp.quizy.matching.domain.user.UserId
+
+interface MatchingQueueRepository {
+    fun enqueue(status: UserStatus)
+
+    fun dequeue(): UserStatus?
+
+    fun saveSet(userId: UserId)
+
+    fun removeSet(userId: UserId)
+
+    fun isValid(userId: UserId): Boolean
+}
