+++ matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/domain/match/MatchingService.kt
@@ -1,0 +1,16 @@
+package com.grepp.quizy.matching.domain.match
+
+import com.grepp.quizy.matching.domain.user.UserId
+import org.springframework.stereotype.Service
+
+@Service
+class MatchingService(
+    private val matchingPoolManager: MatchingPoolManager,
+    private val matchingEventInternalPublisher: MatchingEventInternalPublisher
+) : MatchingUseCase {
+
+    override fun registerWaiting(userId: UserId) {
+        matchingPoolManager.save(userId)
+        matchingEventInternalPublisher.publish(UserWaitingRegisteredEvent(userId))
+    }
+}
