+++ matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/domain/match/MatchingUseCase.kt
@@ -1,0 +1,7 @@
+package com.grepp.quizy.matching.domain.match
+
+import com.grepp.quizy.matching.domain.user.UserId
+
+interface MatchingUseCase {
+    fun registerWaiting(userId: UserId)
+}
