--- matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/game/GameFetcher.kt
@@ -1,17 +1,0 @@
-package com.grepp.quizy.matching.game
-
-import com.grepp.quizy.matching.user.InterestCategory
-import com.grepp.quizy.matching.user.UserId
-import org.springframework.stereotype.Component
-
-interface GameFetcher {
-    fun requestGameRoomId(userIds: List<UserId>, subject: InterestCategory): GameRoomId
-}
-
-// @TODO feignclient로 요청 필요
-@Component
-class GameFetcherImpl : GameFetcher {
-    override fun requestGameRoomId(userIds: List<UserId>, subject: InterestCategory): GameRoomId {
-        return GameRoomId(1L)
-    }
-}
