--- matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/game/GameVO.kt
@@ -1,3 +1,0 @@
-package com.grepp.quizy.matching.game
-
-@JvmInline value class GameRoomId(val value: Long)
