--- matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/match/MatchingConstants.kt
@@ -1,3 +1,0 @@
-package com.grepp.quizy.matching.match
-
-const val MATCHING_K = 5
