--- matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/match/MatchingQueueRepository.kt
@@ -1,15 +1,0 @@
-package com.grepp.quizy.matching.match
-
-import com.grepp.quizy.matching.user.UserId
-
-interface MatchingQueueRepository {
-    fun enqueue(status: UserStatus)
-
-    fun dequeue(): UserStatus?
-
-    fun saveSet(userId: UserId)
-
-    fun removeSet(userId: UserId)
-
-    fun isValid(userId: UserId): Boolean
-}
