--- matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/match/MatchingService.kt
@@ -1,16 +1,0 @@
-package com.grepp.quizy.matching.match
-
-import com.grepp.quizy.matching.user.UserId
-import org.springframework.stereotype.Service
-
-@Service
-class MatchingService(
-    private val matchingPoolManager: MatchingPoolManager,
-    private val matchingEventInternalPublisher: MatchingEventInternalPublisher
-) : MatchingUseCase {
-
-    override fun registerWaiting(userId: UserId) {
-        matchingPoolManager.save(userId)
-        matchingEventInternalPublisher.publish(UserWaitingRegisteredEvent(userId))
-    }
-}
