--- matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/match/MatchingUseCase.kt
@@ -1,7 +1,0 @@
-package com.grepp.quizy.matching.match
-
-import com.grepp.quizy.matching.user.UserId
-
-interface MatchingUseCase {
-    fun registerWaiting(userId: UserId)
-}
