--- matching-service/matching-domain/src/main/kotlin/com/grepp/quizy/matching/user/UserVectorizer.kt
@@ -1,20 +1,0 @@
-package com.grepp.quizy.matching.user
-
-import org.springframework.stereotype.Component
-
-@Component
-class UserVectorizer(private val userFetcher: UserFetcher) {
-
-    fun vectorizeUser(userId: UserId): UserVector {
-        val rating = userFetcher.requestUserRating(userId)
-        val interests = userFetcher.requestUserInterests(userId)
-        val vectorSize = GameRating.entries.size + InterestCategory.entries.size
-
-        return UserVector(
-            FloatArray(vectorSize).apply {
-                this[rating.index] = 1.0f
-                interests.forEach { this[it.index + InterestCategory.VECTOR_START_INDEX] = 1.0f }
-            }
-        )
-    }
-}
