+++ matching-service/matching-infra/src/main/kotlin/com/grepp/quizy/matching/infra/match/pubsub/MatchingRedisPublisher.kt
@@ -1,0 +1,18 @@
+package com.grepp.quizy.matching.infra.match.pubsub
+
+import com.grepp.quizy.matching.match.MatchingEventPublisher
+import com.grepp.quizy.matching.match.PersonalMatchingSucceedEvent
+import org.springframework.data.redis.core.RedisTemplate
+import org.springframework.data.redis.listener.ChannelTopic
+import org.springframework.stereotype.Component
+
+@Component
+class MatchingRedisPublisher(
+    private val matchingMessageTemplate: RedisTemplate<String, PersonalMatchingSucceedEvent>,
+) : MatchingEventPublisher {
+
+    override fun publish(event: PersonalMatchingSucceedEvent) {
+        val topic = ChannelTopic("matching")
+        matchingMessageTemplate.convertAndSend(topic.topic, event)
+    }
+}
